fix: clean up chat screen
